requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
supabase==2.3.0
psycopg2-binary==2.9.9
tqdm==4.66.1
pandas==2.1.4
numpy==1.26.3
rich==13.7.0 
tiktoken
pyyaml